Conflicts:
	CHANGELOG

+++ protected/modules/category/install/migrations/m000000_000000_category_base.php
@@ -1,30 +1,76 @@
 <?php
+/**
+ * FileDocComment
+ * Category install migration
+ * Класс миграций для модуля Category:
+ *
+ * @category YupeMigration
+ * @package  YupeCMS
+ * @author   YupeTeam <team@yupe.ru>
+ * @license  BSD https://raw.github.com/yupe/yupe/master/LICENSE
+ * @link     http://yupe.ru
+ **/
+
+/**
+ * Category install migration
+ * Класс миграций для модуля Category:
+ *
+ * @category YupeMigration
+ * @package  YupeCMS
+ * @author   YupeTeam <team@yupe.ru>
+ * @license  BSD https://raw.github.com/yupe/yupe/master/LICENSE
+ * @link     http://yupe.ru
+ **/
 class m000000_000000_category_base extends CDbMigration
 {
+    /**
+     * Функция настройки и создания таблицы:
+     *
+     * @return nothing
+     **/
     public function safeUp()
     {
         $db = $this->getDbConnection();
-        $tableName = $db->tablePrefix.'category';
-        $this->createTable($tableName, array(
-            'id' => 'pk',
-            'parent_id' => 'integer DEFAULT NULL',
-            'alias' => 'string NOT NULL',
-            'lang' => 'char(2) DEFAULT NULL',
-            'name' => 'string NOT NULL',
-            'image' => 'varchar(300) DEFAULT NULL',
-            'short_description' => 'text',
-            'description' => 'text NOT NULL',
-            'status' => "smallint(1) NOT NULL DEFAULT '1'",
-        ),"ENGINE=InnoDB DEFAULT CHARSET=utf8");
+        $this->createTable(
+            $db->tablePrefix . 'category', array(
+                'id' => 'pk',
+                'parent_id' => 'integer DEFAULT NULL',
+                'alias' => 'string NOT NULL',
+                'lang' => 'char(2) DEFAULT NULL',
+                'name' => 'string NOT NULL',
+                'image' => 'varchar(300) DEFAULT NULL',
+                'short_description' => 'text',
+                'description' => 'text NOT NULL',
+                'status' => "smallint(1) NOT NULL DEFAULT '1'",
+            ), "ENGINE=InnoDB DEFAULT CHARSET=utf8"
+        );
 
-        $this->createIndex("category_alias_uniq",$tableName,"alias,lang", true);
-        $this->createIndex("category_parent_id",$tableName,"parent_id", false);
-        $this->createIndex("category_status",$tableName,"status", false);
+        $this->createIndex($db->tablePrefix . "category_alias_uniq", $db->tablePrefix . 'category', "alias,lang", true);
+        $this->createIndex($db->tablePrefix . "category_parent_id", $db->tablePrefix . 'category', "parent_id", false);
+        $this->createIndex($db->tablePrefix . "category_status", $db->tablePrefix . 'category', "status", false);
     }
  
+    /**
+     * Функция удаления таблицы:
+     *
+     * @return nothing
+     **/
     public function safeDown()
     {
         $db = $this->getDbConnection();
-        $this->dropTable($db->tablePrefix.'category');
+
+        /**
+         * Убиваем внешние ключи, индексы и таблицу - settings
+         * @todo найти как проверять существование индексов, что бы их подчищать (на абстрактном уровне без привязки к типу БД):
+         **/
+
+        if ($db->schema->getTable($db->tablePrefix . 'category') !== null) {
+            /*
+            $this->dropIndex($db->tablePrefix . "category_alias_uniq", $db->tablePrefix . 'category');
+            $this->dropIndex($db->tablePrefix . "category_parent_id", $db->tablePrefix . 'category');
+            $this->dropIndex($db->tablePrefix . "category_status", $db->tablePrefix . 'category');
+            */
+            $this->dropTable($db->tablePrefix . 'category');
+        }
     }
 }
